requirements.txt
pytest-playwright>=0.5.1
pytest-html>=4.1.1
pytest-json-report>=1.5.0
python-dotenv~=1.0.1